requirements.txt
aiogram==3.7.0
SQLAlchemy[asyncio]==2.0.34
aiosqlite==0.19.0
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.2
yt-dlp>=2025.11.12
youtube-search-python>=1.6.6
requests==2.31.0

orjson==3.10.7
rapidfuzz==3.9.7
# Optional for voice search:
vosk==0.3.45
# For music note extraction:
librosa==0.10.1
numpy==1.24.3